% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cawCV-package.R
\docType{package}
\name{cawCV-package}
\alias{cawCV}
\alias{cawCV-package}
\title{cawCV: Cluster-Aware Cross-Validation Partitioning}
\description{
Cluster-aware cross-validation partitioning and pre-validation diagnostics
for clinical prediction models with repeated-measures predictors and
subject-level binary outcomes.
}
\author{
\strong{Maintainer}: Joseph Hagan \email{jlhagan@bcm.edu} (\href{https://orcid.org/0000-0002-3285-2131}{ORCID})

}
\keyword{internal}
