% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caw_partition.R
\name{n_folds}
\alias{n_folds}
\title{Number of folds in a partition}
\usage{
n_folds(partition)
}
\arguments{
\item{partition}{A \code{caw_partition} object.}
}
\value{
An integer giving the number of folds.
}
\description{
Number of folds in a partition
}
\examples{
p <- partition_subject_kfold(rep(1:12, each = 5), k = 4, seed = 1)
n_folds(p)
}
