% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caw_partition.R
\name{print.caw_partition}
\alias{print.caw_partition}
\title{Print a cross-validation partition}
\usage{
\method{print}{caw_partition}(x, ...)
}
\arguments{
\item{x}{A \code{caw_partition} object.}

\item{...}{Ignored. Present for S3 method consistency.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a cross-validation partition
}
